+++ docs/changelog.md
@@ -14,3 +14,4 @@
 - Argument '--readfile' is now '--read-file'
 - You can now run 'ng++' to execute the program instead of only 'numguesser++'
 - Implemented the '-n' argument from the original version of numguesser, which prevents the correct number from being displayed after a loss.
+- Implemented a '--debug' argument that displays debug information about program variables and terminates. The same information is also displayed when terminating the program with Ctrl-C